rename examples, move models to a seperate file, add RankFilterPruner base class
